+++ src/pixi-node/index.ts
@@ -1,4 +1,4 @@
 export * from "./NodeGPUCanvas.ts";
 export * from "./NodeDOMAdapter.ts";
-export * from "./NodeTextCanvas.ts";
+export * from "./NodeCanvas.ts";
 export * from "./createPixiRenderer.ts";